Extract stuff to remove to isolated file
